requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
aiosqlite==0.21.0
    # via prefect
alembic==1.17.0
    # via prefect
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   prefect
    #   starlette
apprise==1.9.5
    # via prefect
asgi-lifespan==2.1.0
    # via prefect
async-timeout==5.0.1
    # via
    #   asyncpg
    #   python-socks
asyncpg==0.30.0
    # via prefect
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
cachetools==5.5.2
    # via
    #   google-auth
    #   prefect
certifi==2025.10.5
    # via
    #   apprise
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.1.8
    # via
    #   apprise
    #   prefect
    #   typer
    #   uvicorn
cloudpickle==3.1.1
    # via prefect
colorama==0.4.6
    # via griffe
confluent-kafka==2.12.0
    # via -r requirements.in
coolname==2.2.0
    # via prefect
cryptography==46.0.3
    # via prefect
dateparser==1.2.2
    # via prefect
docker==7.1.0
    # via
    #   -r requirements.in
    #   prefect
exceptiongroup==1.3.0
    # via
    #   anyio
    #   prefect
fastapi==0.119.1
    # via
    #   -r requirements.in
    #   prefect
fsspec==2025.9.0
    # via prefect
google-api-core==2.26.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery==3.28.0
    # via -r requirements.in
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.4.1
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
graphviz==0.21
    # via prefect
greenlet==3.2.4
    # via sqlalchemy
griffe==1.14.0
    # via prefect
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via
    #   httpx
    #   prefect
httpx==0.28.1
    # via prefect
humanize==4.14.0
    # via
    #   jinja2-humanize-extension
    #   prefect
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via
    #   jinja2-humanize-extension
    #   prefect
jinja2-humanize-extension==0.4.0
    # via prefect
jsonpatch==1.33
    # via prefect
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via prefect
jsonschema-specifications==2025.9.1
    # via jsonschema
kafka-python==2.2.15
    # via -r requirements.in
mako==1.3.10
    # via alembic
markdown==3.9
    # via apprise
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.6
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
opentelemetry-api==1.38.0
    # via prefect
orjson==3.11.3
    # via prefect
packaging==24.2
    # via prefect
pandas==2.3.3
    # via -r requirements.in
pathspec==0.12.1
    # via prefect
pendulum==3.1.0
    # via prefect
prefect==3.2.15
    # via
    #   -r requirements.in
    #   prefect-shell
prefect-shell==0.3.1
    # via -r requirements.in
prometheus-client==0.23.1
    # via prefect
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
py4j==0.10.9.9
    # via pyspark
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   fastapi
    #   prefect
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.4
    # via
    #   prefect
    #   pydantic
pydantic-extra-types==2.10.6
    # via prefect
pydantic-settings==2.11.0
    # via prefect
pygments==2.19.2
    # via rich
pyspark==4.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   pandas
    #   pendulum
    #   prefect
python-dotenv==1.1.1
    # via pydantic-settings
python-slugify==8.0.4
    # via prefect
python-socks==2.8.0
    # via prefect
pytz==2025.2
    # via
    #   dateparser
    #   pandas
    #   prefect
pyyaml==6.0.3
    # via
    #   apprise
    #   prefect
readchar==4.2.1
    # via prefect
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.10.23
    # via dateparser
requests==2.32.5
    # via
    #   apprise
    #   docker
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via apprise
rfc3339-validator==0.1.4
    # via prefect
rich==13.9.4
    # via
    #   prefect
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruamel-yaml==0.18.15
    # via prefect
ruamel-yaml-clib==0.2.14
    # via ruamel-yaml
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   asgi-lifespan
    #   prefect
sqlalchemy==2.0.44
    # via
    #   alembic
    #   prefect
starlette==0.48.0
    # via fastapi
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via prefect
tomli==2.3.0
    # via
    #   alembic
    #   confluent-kafka
typer==0.15.4
    # via prefect
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   alembic
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   opentelemetry-api
    #   prefect
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   referencing
    #   rich
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via
    #   pandas
    #   pendulum
tzlocal==5.3.1
    # via dateparser
ujson==5.11.0
    # via prefect
urllib3==2.5.0
    # via
    #   docker
    #   requests
uv==0.9.5
    # via prefect
uvicorn==0.38.0
    # via
    #   -r requirements.in
    #   prefect
websockets==15.0.1
    # via prefect
zipp==3.23.0
    # via importlib-metadata
